pyproject.toml
[project]
name = "feishu-kirocli-bot"
version = "0.1.0"
description = "Feishu (Lark) bot powered by Kiro CLI via ACP protocol"
requires-python = ">=3.9"
dependencies = [
    "lark-oapi>=1.5.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
feishu-kirocli-bot = "bridge:main"
